requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyJWT